Fixes dependency errors
