+++ manage.py
@@ -13,7 +13,7 @@ ENV = os.environ.get("ENV", "test")
 ENV = ENV.lower()
 
 if ENV not in SUPPORTED_ENVS:
-    raise Exception(f"Unsupported environment: {ENV}")
+    raise EnvironmentError(f"Unsupported environment: {ENV}")
 
 if __name__ == "__main__":
     os.environ.setdefault("DJANGO_SETTINGS_MODULE", SETTINGS_MODULES[ENV])
